requirements.txt
langchain
langchain-openai
langchain-community
faiss-cpu
huggingface_hub
gradio
pdfplumber
python-docx
python-dotenv
fastapi
uvicorn
sentence-transformers
